+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - ESP32 Increase number of switch GPIOs from 8 to 28
 - ESP32 Increase number of interlock groups from 4 to 14
 - Increase number of button GPIOs from 4 to 8
+- Preview of Berry language for Tasmota32
 
 ## [9.2.0.5] 20210205
 ### Changed